spec(2.3) Sprint A Step 6: D7+D8 catversion bump + msg_types SRF/view (atomic)
Atomic catalog change commit per L46 (catversion bump must cover any
new SRF/view exposed via pg_proc.dat).  Combines:

D7  catversion 202605200 -> 202605210
    - src/include/catalog/catversion.h (+ block comment per L46)
    - Reason: wire-protocol pivot from spec-2.2 24-byte ClusterMsgHeader
      to spec-2.3 36-byte ClusterICEnvelope (incompatible WAL/IC bytes:
      a 202605200 binary cannot open a 202605210 datadir + vice versa)
      AND new pg_cluster_ic_msg_types view + cluster_get_ic_msg_types
      SRF added to pg_proc.dat.
    - L46 三维 grep audit: cluster_tap (075/068/069) + cluster_regress
      (ic_tcp_smoke) all use lower-bound regex (>= 202605200), so
      bumping to 202605210 does not break them.  Spec-2.3 body
      already uses 202605210 throughout.

D8  cluster_get_ic_msg_types SRF + pg_cluster_ic_msg_types view
    - pg_proc.dat: oid 8915 (next free after spec-2.2 D9's 8914)
      record SRF returning (msg_type int4, name text,
      allowed_producer_mask int8, broadcast_ok bool,
      handler_present bool); STABLE volatility (process-local
      dispatch_table populated at postmaster phase 1).
    - system_views.sql: pg_cluster_ic_msg_types view + REVOKE +
      GRANT SELECT TO PUBLIC pattern (mirrors spec-2.2 D9
      pg_cluster_ic_peers).
    - cluster_ic_msg_types_srf.c (NEW; ~80 LOC; --enable-cluster only):
      iterates cluster_ic_get_msg_type_info() public accessor; lives
      in own file so cluster_unit standalone tests don't hit
      InitMaterializedSRF / cstring_to_text / tuplestore_putvalues
      unresolved symbols (mirrors spec-2.2 cluster_get_ic_peers
      living in cluster_ic_tier1.c).
    - cluster_ic.c: PG_FUNCTION_INFO_V1(cluster_get_ic_msg_types)
      always-linked + disable-cluster ERRCODE_FEATURE_NOT_SUPPORTED
      stub at file bottom (mirrors cluster_get_ic_peers pattern).
    - src/backend/cluster/Makefile: add cluster_ic_msg_types_srf.o
      to OBJS (USE_PGRAC_CLUSTER branch).
    - src/test/regress/expected/rules.out: add view dump entry
      between pg_get_backend_memory_contexts and pg_cluster_ic_peers.

Cleanup:
    - cluster_ic.h: remove duplicate
      #define PGRAC_IC_ENVELOPE_VERSION_V1 ((uint16)1) -- the
      authoritative definition is now ((uint8)1) in
      cluster_ic_envelope.h (HELLO struct still uses 2-byte
      slot via implicit promotion at use site).  Resolves
      -Wmacro-redefined.
    - cluster_ic_router.c: cast `info->msg_type` to int when
      comparing against CLUSTER_IC_MSG_TYPE_MAX (256) -- silences
      -Wtautological-constant-out-of-range-compare while keeping
      the defensive bound check structure.

Verification surfaces (local):
    - cluster_unit: 33/33 binaries pass; total 250+ sub-tests
    - cluster_regress: 4/4 (cluster_smoke, undo_tablespace,
      conf_multi_node_activation, ic_tcp_smoke) pass
    - PG 219/219 regression pass (rules.out updated)
    - cluster_tap 075-080: deferred to CI (local mac lacks IPC::Run)
    - clang-format-18 (CI version): 0 violations
    - cppcheck: 0 new findings introduced by spec-2.3 files
      (42 baseline drift in cluster_unit test files unchanged)

Spec authority: pgrac:specs/spec-2.3-...md frozen v0.2 §1.2 D7/D8.
Lessons SSOT alignment: L46 (catversion atomic batch), L66 (CI green
incl. remote workflows -- gated by Step 5 success run 25528313287).

Spec: spec-2.3-envelope-abi-ratify-transport-agnostic-api.md

+++ src/backend/cluster/cluster_ic_router.c
@@ -100,7 +100,7 @@ cluster_ic_register_msg_type(const ClusterICMsgTypeInfo *info)
 		ereport(FATAL, (errcode(ERRCODE_INTERNAL_ERROR),
 						errmsg("cluster_ic_register_msg_type called with NULL info")));
 
-	if (info->msg_type == 0 || info->msg_type >= CLUSTER_IC_MSG_TYPE_MAX)
+	if (info->msg_type == 0 || (int)info->msg_type >= CLUSTER_IC_MSG_TYPE_MAX)
 		ereport(FATAL, (errcode(ERRCODE_INTERNAL_ERROR),
 						errmsg("cluster_ic_register_msg_type: msg_type %u out of range "
 							   "[1, %d)",
@@ -151,7 +151,7 @@ cluster_ic_send_envelope(uint8 msg_type, int32 dest_node_id, const void *payload
 	const ClusterICMsgTypeInfo *info;
 	ClusterICEnvelope env;
 
-	if (msg_type == 0 || msg_type >= CLUSTER_IC_MSG_TYPE_MAX)
+	if (msg_type == 0 || (int)msg_type >= CLUSTER_IC_MSG_TYPE_MAX)
 		ereport(ERROR, (errcode(ERRCODE_INVALID_PARAMETER_VALUE),
 						errmsg("cluster_ic_send_envelope: msg_type %u out of range", msg_type)));
 
@@ -227,7 +227,7 @@ cluster_ic_dispatch_envelope(const ClusterICEnvelope *env, const void *payload)
 	if (env == NULL)
 		return false;
 
-	if (env->msg_type == 0 || env->msg_type >= CLUSTER_IC_MSG_TYPE_MAX)
+	if (env->msg_type == 0 || (int)env->msg_type >= CLUSTER_IC_MSG_TYPE_MAX)
 		return false; /* peer sent malformed msg_type; caller close peer */
 
 	info = &dispatch_table[env->msg_type];
@@ -277,7 +277,7 @@ cluster_ic_dispatch_envelope(const ClusterICEnvelope *env, const void *payload)
 const ClusterICMsgTypeInfo *
 cluster_ic_get_msg_type_info(uint8 msg_type)
 {
-	if (msg_type == 0 || msg_type >= CLUSTER_IC_MSG_TYPE_MAX)
+	if (msg_type == 0 || (int)msg_type >= CLUSTER_IC_MSG_TYPE_MAX)
 		return NULL;
 	if (!slot_registered(&dispatch_table[msg_type]))
 		return NULL;
@@ -297,5 +297,6 @@ cluster_ic_router_count_registered(void)
 	}
 	return count;
 }
+
 
 #endif /* USE_PGRAC_CLUSTER */
